VideoPlayerActivity: missing cast

+++ vlc-android/src/org/videolan/vlc/gui/video/VideoPlayerActivity.java
@@ -1205,11 +1205,11 @@ public class VideoPlayerActivity extends Activity implements IVideoPlayer {
             //TODO : Volume action when a secondary display is connected
             if (coef > 2 && mPresentation == null) {
                 // Volume (Up or Down - Right side)
-                if (!mEnableBrightnessGesture || mTouchX > (screen.widthPixels / 2)){
+                if (!mEnableBrightnessGesture || (int)mTouchX > (screen.widthPixels / 2)){
                     doVolumeTouch(y_changed);
                 }
                 // Brightness (Up or Down - Left side)
-                if (mEnableBrightnessGesture && mTouchX < (screen.widthPixels / 2)){
+                if (mEnableBrightnessGesture && (int)mTouchX < (screen.widthPixels / 2)){
                     doBrightnessTouch(y_changed);
                 }
                 // Extend the overlay for a little while, so that it doesn't
